BibEdit: Integration with the Bibcirculation module
If an appropriate parameter is set in the Invenio configuration, a Bibcirculation
panel is displayed inside BibEdit, allowing to see the number of physical copies
and to easily jump to an appropriate bibCirculation page allowing to edit these.

The visibility of the panel is also subject to the classification of a record.
It is displayed only for record saved in the database and belonging to a collection
for which the holding tab is being displayed